Finish conversion to formatter class
